Update rss.php
The RSS feeds in both 1.0.6 and 1.1.0 include the price in the title, but when the price is obtained in rss.php, the print_money function isn't set correctly; it adds a hyperlink to the currency converter to the price, and since the title is in CDATA tags, it's output as plain text. To fix it, you simply need to add the “false” for the link parameter to the function call in rss.php.