Update rss.php
These modifications fix a number of issues with the RSS feeds for WeBid. 

The default rss.php and rss.tpl use the site name for the feed title and description. This uses the meta description for the description instead of repeating the title.

You also add the Content Type via a header() call so that the RSS feeds validate.