Fixes closing item problem with one or more bids
Items were closing with 1 or more bids & not reaching the reserve price and not making it to the yourauctions_c.php page. They just closed and only the admin could see them. This update fixes that issue and now items with more than one bid end up listed on the close items page.