Added working checkboxes
Updated code for the checkboxes to appear for the relist and 'sell under' reserved price to highest bidder. Checkboxes for items with bids show up for the relist & sell under. Checkboxes for items with no bid show up under the re-list and delete. I am working to get the sell under code to move the item to the sold page. It does send an email to the highest bidder telling them they won and the highest bidder can buy the item.