Cosmetic alterations for featured items
To save confusion I have added headers to the featured items on the
homepage as well as the browse page, also added a new header ALL ITEMS
on the browse page.

index.php has had code added to only display the featured part if
featured items exist.

Many posts on the forum where people and their users are confused