Dahlsvarehus's captcha update
Dahlsvarehus has an excellen mod here. I want to give Dahlsvarehus full credit for this. This update on the captcha stops bots from registering on our webid sites. My site gets a bot trying to register about every 5 minutes and none have been successful since I added this update to my old captcha. The bots had no problem registering when I was using the old basic captcha. I want to vote for this to be the mod of the month.