Fix the reserve price bug
If the reserve price is selected before the Buy Now Only was selected, it caused an error and even if the reserve price was at zero. The error read "The Reserve price you entered is not correct". The error reported makes no sense and it really doesn't tell users to switch the reserve to  "no". This bug did not let the user proceed to sell the item until the reserve was switched to "no" if the user wanted to sell an item using the "buy now only. The fix here fixes that bug and will switch the reserve to the correct selection when the buy now is selected "yes" or "no".