Fix for issue #92
this fix will check to see if you have defined a CHARSET via the
config.inc.php if not it does not use it to connect to database as it is
an OPTIONAL variable for the connect command.

also moved connect to database to be above the other commands as they
require it
tested on existing database works fine.