Update faqscategories.tpl
On this page we have 2 submit buttons. Which one is to stay and which one is to go? I have noticed the <button> has been  replacing the normal submit input keeping the posted var consistent and not changing if another language is used while the actual text can alter. Totaly agree