More Various Changes
- Added function filter_date in global functions.
This can be used on other pages to validate and format date input.
- Modified function FormatTimeStamp in dates.php to work on nermal
calender format and to also rectify double swapping of day and month as
per thread in forum.
This function is only used in the accounts.php only at the momentso
makes sense to modify it and have default "-" as the devider.
- Increments.php moved the end $db->query() from the end of the if
statments to each query as there was a possibility of the script
reaching the call without entering the if's, so no parameters would be
set.
- Editauction.php asdded isset check. Also in the load_gallery function
removed pointless $K which has no initial value and the array self
assigns the keys anyway.
- Viewfilters.php added missing $current_page.
- Editbanner.php corrected small error in name preventing the replacment
of an existing banner to complete the image transfer