New Settings Table and Shipping Option
This is an initial commit as part of a set of changes to how the
settings table will work.

A new settings table has to be added to the database manually for now:
CREATE TABLE `webid`.`webid_settingsv2` ( `fieldname` VARCHAR(30) NOT
NULL ,`fieldtype` VARCHAR(10) NOT NULL , `value` VARCHAR(255) NOT NULL
,`modifieddate` INT(11) NOT NULL , `modifiedby` INT(32) NOT NULL ,
PRIMARY KEY(`fieldname`))

And a value needs to be inserted using the following (replacing the
database name and table name as appropriate to your environment):
INSERT INTO `webid`.`webid_settingsv2` (`fieldname`, `fieldtype`,
`value`, `modifieddate`, `modifiedby`) VALUES ('shipping', 'boolean',
'1', '0', '0')

Replace the value of 1 with a 0 if you want to disable shipping.