Removed jQuery gallery lightbox function
I was getting this error in Google Chrome's console on every single
page:
Uncaught TypeError: $(...).lightBox is not a function

This function should only work on the item page as that is the only
location in the theme where there is an id of gallery with anchor tags
within.

It has now been moved to the item.tpl file and should only be loaded if
there is a gallery for that particular auction item.