Presetting some variables and updating the coding
Eliminated all the global in the functions and just left 1 in the __construct() function so that you can just use $this-> in all the functions for the $DBPrefix, $db, $system, $user and these variables are using private so they can only be called and use in side it own class. I also cleaned up the paypal_validate() function so that it use the preset $this->paypalAddress variable with less "if" statements to handle live and sandbox address with 1 "if" statement to check if the site using SSL or not so that it opens the correct connection to paypal. i also cleaned up the $params arrays for the database querys. I also put the callback_process() function on private so that it cannot be called out side the class and set all the $this->callback_process to self::callback_process() so that the function can be used.

Sorry but i am not good at the proper terms i am just good at know how to do it.