Update to .htaccess
I think it's time I share my .htaccess codes since I've tested these out for the last couple of years. The browser user-agents can be omitted if people are scared to use them but, blocking old browsers is some of my best defenses. If the added browser blocks cause problems for some reason  then they can be removed down the road. The blocks are for a few Firefox's below 10 and the 10's& 20's. Then the blocks are for the Chrome 10s, 20s, and 30s. No need to block the 40s as of now.