Upload Fix
I commented out the allowed_mime_types section until we fix it. It's not allowing photos to be uploaded. 

In the 2nd part I removed the .part because the files were being named with a .part extension and the thumbnail preview wasn't finding it.

This all still needs to be worked on but, it will fix the image uploads issue in the meantime.