Removed B_HIGHLIGHTED
This was causing Unknown error type: [8] Undefined index: B_HIGHLIGHTED when running search, advanced search when finding items and also when clicking a category with items.. It's not in any of the old scripts that I checked. For some reason it was added to the modern theme. Removing it will stop all of the errors that pop up in the error log.