Time adjustment
The removal of the parts of the codes will allow the user's newly submitted items to be inserted into the database at their correct timezone instead of UTC's timezone. For me the UTC timezone is 4 hours ahead of my EST timezone so, it caused my newly submitted auctions to be pending for 4 hours. There was also two time conversions on each of the lines adjusted. For instance the $a_starts & $a_ends has convertToDatetime in it's definition. Timezone will adjust correctly with whatever timezone the user chooses with the  code this way. The other way, timezones won't adjust at all on newly submitted items.