Ends within counter fix
When viewing an item the "Ends within" counter was displaying an incorrect value. The counter was not counting when the browser tab was inactive in Chrome or on mobile.

This fix changes the counter logic to use the current time to calculate the time remaining, ensuring that the counter is always accurate. Even when the browser tab has been inactive.